Fixed most of the major stickies bugs and added numerous options, including always-on-top notes, custom system tray icons, start Stickies with Windows, etc.
Made the behavior appropriate when the user tries to open Stickies multiple
times.

Got rid of version checking in my attempt to use ClickOnce as a distribution
platform.  Given that it doesn't work properly when Firefox is the default
browser, I am skeptical I will do this in the long-run, but haven't
decided yet.

Refactored a bunch of the "keep the notes on the screen" logic so it could
be reused for the About form and the Preferences form.

Added a global Windows hotkey for creating new notes (Ctrl-Shift-N). Notes
are now created underneath the mouse cursor by default.