Added padding around the note text and made the note border smaller, which is more aesthetically pleasing.
Fixed http://code.google.com/p/stickies-windows/issues/detail?id=10 by
incorporating the bottom border height in the rolled up height, which adds
the additional pixel we need for taller fonts.

The only issue with the extra padding around the text box is that the text
box scroll bar is offset from the border by a few pixels, which may look
weird.  I don't notice it, and scrolling is not the common case, so I am
leaving it for now.